Added 'type' to filters; moved filter components to ButtonGroup and MultiSelect; renamed custom NextUI Select components
