+++ src/components/FilterView.tsx
@@ -12,54 +12,63 @@ const filters = [
     display: "Priority Level",
     options: ["Low", "Medium", "High"],
     tooltip: "For information on how this is calculated, see the About page",
+    type: "buttonGroup",
   },
   {
     property: "parcel_type",
     display: "Parcel Type",
     options: ["Land", "Building"],
     tooltip: "Parcel type from City of Philadelphia data",
+    type: "buttonGroup",
   },
   {
     property: "access_process",
     display: "Access Process",
     options: ["Buy Property", "Land Bank", "Private Land Use Agreement"],
     tooltip: "For information on what these mean, see the Get Access page",
+    type: "buttonGroup",
   },
   {
     property: "neighborhood",
     display: "Neighborhoods",
     options: neighborhoods,
     tooltip: "Neighborhood mapping from OpenDataPhilly by Element 84 (formerly Azavea)",
+    type: "multiSelect",
   },
-  // {
-  //   property: "rco_info",
-  //   display: "Community Organizations",
-  //   options: rcos,
-  //   tooltip: "RCO mapping from City of Philadelphia data",
-  // },
+  {
+    property: "rco_names",
+    display: "Community Organizations",
+    options: rcos,
+    tooltip: "RCO mapping from City of Philadelphia data",
+    type: "multiSelect",
+  },
   {
     property: "tactical_urbanism",
     display: "Tactical Urbanism",
     options: ["Yes", "No"],
     tooltip: "For an explanation of this, see the Get Access page",
+    type: "buttonGroup",
   },
   {
     property: "conservatorship",
     display: "Conservatorship Eligible",
     options: ["Yes", "No"],
     tooltip: "For an explanation of this, see the Get Access page",
+    type: "buttonGroup",
   },
   {
     property: "side_yard_eligible",
     display: "Side Yard Eligible",
     options: ["Yes", "No"],
     tooltip: "For an explanation of this, see the Get Access page",
+    type: "buttonGroup",
   },
   {
     property: "llc_owner",
     display: "LLC Owner",
     options: ["Yes", "No"],
     tooltip: "For an explanation of this, see the Get Access page",
+    type: "buttonGroup",
   },
 ];
 
@@ -77,13 +86,14 @@ const FilterView: FC<FilterViewProps> = ({ updateCurrentView }) => {
         startContent={<PiX />}
         onPress={() => updateCurrentView("filter")}
       />
-      {filters.map(({ property, display, options, tooltip }) => (
+      {filters.map(({ property, display, options, tooltip, type }) => (
         <DimensionFilter
           key={property}
           property={property}
           options={options}
           display={display}
           tooltip={tooltip}
+          type={type}
         />
       ))}
     </div>
