+++ src/components/Filters/DimensionFilter.tsx
@@ -1,17 +1,18 @@
 "use client";
 
 import React, { useState, FC } from "react";
-import { Button, Chip, Tooltip } from "@nextui-org/react";
+import { Tooltip } from "@nextui-org/react";
 import { useFilter } from "@/context/FilterContext";
-import { Check, Info, X } from "@phosphor-icons/react";
-import { rcos_detail } from "./FilterOptions"
-import { FilterChip, MultiSelect, MultiSelectItem, BlankSelectorIcon } from "./MultiSelectVariants"
+import { Info } from "@phosphor-icons/react";
+import ButtonGroup from "./ButtonGroup";
+import MultiSelect from "./MultiSelect";
 
 type DimensionFilterProps = {
   property: string;
   display: string;
   options: string[];
   tooltip: string;
+  type: string;
 };
 
 const DimensionFilter: FC<DimensionFilterProps> = ({
@@ -19,6 +20,7 @@ const DimensionFilter: FC<DimensionFilterProps> = ({
   display,
   options,
   tooltip,
+  type,
 }) => {
   const { dispatch, appFilter } = useFilter();
   const [selectedKeys, setSelectedKeys] = useState<string[]>(
@@ -40,10 +42,6 @@ const DimensionFilter: FC<DimensionFilterProps> = ({
   const handleSelectionChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     const newMultiSelect: string[] = e.target.value.split(",")
     setSelectedKeys(newMultiSelect);
-    // If the rco filter includes more than the rco name, 'newDimensions' should be passed to dispatch's 'dimensions' instead of newMultiSelect.  Also in that case, 'newDimensions' should also be added to 'toggleDimension' to filter out rco dimensions.
-    // const newDimensions = (property === 'neighborhood') ? newMultiSelect : newMultiSelect.map((rco) => {
-    //   return rcos_detail[rco]
-    // })
     dispatch({
       type: "SET_DIMENSIONS",
       property,
@@ -51,68 +49,7 @@ const DimensionFilter: FC<DimensionFilterProps> = ({
     });
   }
 
-  if (property === "neighborhood" || property === "rco_info") {
-  
-    const multiSelectOptions = options.filter((option) => {
-      if (!selectedKeys.includes(option)) {
-        return option
-      }
-    })
-
-    return (
-      <div className="pb-6">
-        <div className="flex items-center mb-2">
-          <div className="flex items-center">
-            <h2 className="heading-lg">{display}</h2>
-            <Tooltip content={tooltip} placement="top" showArrow color="primary">
-              <Info
-                alt="More Info"
-                className="h-5 w-9 text-gray-500 pl-2 pr-2 cursor-pointer"
-                tabIndex={0}
-              />
-            </Tooltip>
-          </div>
-        </div>
-        <div className="space-x-2 min-h-[33.5px]">
-          <MultiSelect
-            aria-label={display}
-            items={options}
-            variant="flat"
-            size="md"
-            radius="md"
-            isMultiline={true}
-            selectionMode="multiple"
-            placeholder="Select options..."
-            selectorIcon={<BlankSelectorIcon />}
-            selectedKeys={selectedKeys}
-            renderValue={() => {
-              return (
-                <div className="flex flex-wrap gap-y-2">
-                  {selectedKeys.map((option, index) => (
-                    <FilterChip key={index} classNames={{base:"multiSelectChip"}} endContent={<X />} onClose={() => toggleDimension(option)}>{option}</FilterChip>
-                  ))}
-                </div>
-              )
-            }}
-            onChange={handleSelectionChange}
-          >
-            {multiSelectOptions.map((option) => (
-              <MultiSelectItem 
-                key={option} 
-                value={option}
-                classNames={{base:"multiSelectItem"}}
-                selectedIcon={<BlankSelectorIcon />}
-                shouldHighlightOnFocus={false}
-              >
-                {option}
-              </MultiSelectItem>
-            ))}
-          </MultiSelect>
-        </div>
-        {}
-      </div>
-    )
-  } 
+  const filter = (type === "buttonGroup") ? <ButtonGroup options={options} selectedKeys={selectedKeys} toggleDimension={toggleDimension} /> : <MultiSelect display={display} options={options} selectedKeys={selectedKeys} toggleDimension={toggleDimension} handleSelectionChange={handleSelectionChange}/>;
 
   return (
     <div className="pb-6">
@@ -128,29 +65,7 @@ const DimensionFilter: FC<DimensionFilterProps> = ({
           </Tooltip>
         </div>
       </div>
-      <div className="space-x-2 min-h-[33.5px]">
-        {options.map((option, index) => (
-          <Button
-            key={index}
-            disableAnimation
-            onPress={() => toggleDimension(option)}
-            size="sm"
-            color={selectedKeys.includes(option) ? "success" : "default"}
-            className={
-              selectedKeys.includes(option) ? "tagSelected" : "tagDefault"
-            }
-            radius="full"
-            aria-pressed={selectedKeys.includes(option)}
-            startContent={
-              selectedKeys.includes(option) ? (
-                <Check className="w-3 w-3.5 max-h-6" />
-              ) : undefined
-            }
-          >
-            {option}
-          </Button>
-        ))}
-      </div>
+      {filter}
     </div>
   );
 };
